+++ src/main/java/com/ureca/snac/infra/TossPaymentProperties.java
@@ -1,4 +1,4 @@
-package com.ureca.snac.payments;
+package com.ureca.snac.infra;
 
 import lombok.Getter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
